household-accounting-analysis-agent/src/agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=0.1.0
duckdb>=1.0.0
pandas>=2.0.0
boto3>=1.34.0
